fix: remove polling for owned games
And instead, fetch explicitly. this means state accross tabs wont sync but that
is fine at least for now.
